Refactor contracts UI into components
Extract contract-related UI into dedicated components and helpers for reuse and clarity. Added: src/components/client/contracts/contractsHelpers.js (status/pay config, formatters, shared CSS), ContractsBadges.jsx (Toast, Skeleton, StatusBadge, Pagination, ModalWrap, ModalHeader), ContractCard.jsx, ContractDetailModal.jsx, and ContractPaymentsModal.jsx (loads payments/summary from /api/payments/contract/:id and handles loading/empty/overdue states). Removed the large inline UI/utility code from MyContracts.jsx and replaced it with the new modular components to improve maintainability and reduce duplication.
